refactor: extract getSessionUrl into a single shared helper
